+++ netbox/circuits/api/views.py
@@ -29,7 +29,7 @@ class CircuitsFieldChoicesViewSet(FieldChoicesViewSet):
 #
 
 class ProviderViewSet(CustomFieldModelViewSet):
-    queryset = Provider.objects.all()
+    queryset = Provider.objects.prefetch_related('tags')
     serializer_class = serializers.ProviderSerializer
     filter_class = filters.ProviderFilter
 
@@ -59,7 +59,7 @@ class CircuitTypeViewSet(ModelViewSet):
 #
 
 class CircuitViewSet(CustomFieldModelViewSet):
-    queryset = Circuit.objects.select_related('type', 'tenant', 'provider')
+    queryset = Circuit.objects.select_related('type', 'tenant', 'provider').prefetch_related('tags')
     serializer_class = serializers.CircuitSerializer
     filter_class = filters.CircuitFilter