+++ netbox/extras/models.py
@@ -700,8 +700,21 @@ class ConfigContext(models.Model):
     def get_absolute_url(self):
         return reverse('extras:configcontext', kwargs={'pk': self.pk})
 
+    def clean(self):
+
+        # Verify that JSON data is provided as an object
+        if type(self.data) is not dict:
+            raise ValidationError(
+                {'data': 'JSON data must be in object form. Example: {"foo": 123}'}
+            )
+
 
 class ConfigContextModel(models.Model):
+
+    local_context_data = JSONField(
+        blank=True,
+        null=True,
+    )
 
     class Meta:
         abstract = True
@@ -715,6 +728,10 @@ class ConfigContextModel(models.Model):
         data = OrderedDict()
         for context in ConfigContext.objects.get_for_object(self):
             data.update(context.data)
+
+        # If the object has local config context data defined, that data overwrites all rendered data
+        if self.local_context_data is not None:
+            data.update(self.local_context_data)
 
         return data